Move retries to independent package
