+++ autorest/adal/token.go
@@ -31,6 +31,7 @@ import (
 	"time"
 
 	"github.com/Azure/go-autorest/autorest/date"
+	"github.com/Azure/go-autorest/autorest/retriablerequest"
 	"github.com/dgrijalva/jwt-go"
 )
 
@@ -57,12 +58,6 @@ const (
 
 	// imdsEndpoint is the well known endpoint for getting MSI authentications tokens
 	imdsEndpoint = "http://169.254.169.254/metadata/identity/oauth2/token"
-
-	// DefaultRetryAttempts is number of attempts for retry status codes (5xx).
-	defaultRetryAttempts = 5
-
-	// DefaultRetryDuration is the duration to wait between retries.
-	defaultRetryDuration = 1 * time.Second
 )
 
 // OAuthTokenProvider is an interface which should be implemented by an access token retriever
@@ -626,7 +621,7 @@ func (spt *ServicePrincipalToken) refreshInternal(resource string) error {
 		req.Header.Set(metadataHeader, "true")
 	}
 
-	resp, err := retry(req, defaultRetryDuration, defaultRetryAttempts)
+	resp, err := retriablerequest.Retry(req, retriablerequest.DefaultRetryDuration, retriablerequest.DefaultRetryAttempts)
 	if err != nil {
 		return fmt.Errorf("adal: Failed to execute the refresh request. Error = '%v'", err)
 	}
